requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.7.0
orjson==3.10.12
python-docx==1.1.2
reportlab==4.2.5
typing-extensions==4.15.0
tzdata==2025.1
